Limited range of flying units.
Units being able to fly anywhere totally ruins AI performance.  It also
ruins the tactical nature of the game because you can't defend ranged
units against them.  We're going to try limiting their range to see if
that helps.

The AI also has a hard time with a hopeless battle situation, especially
because the units aren't balanced right now.  The score needs to
accurately reflect the likelihood that one side is winning.