Attempt #1 at improving AI performance.
Storing units in the game state naively by entity ID led to a large
amount of wasted space.  Lookups were fast but the AI had to copy a 100+
element vector with only a dozen active units in it.  This first attempt
uses a hash table instead and seems to be faster.